requirements.txt
streamlit
whisper
moviepy
tempfile
google.generativeai
youtube_transcript_api